Changed the output build path to be non-global
EXECUTABLE_OUTPUT_PATH and LIBRARY_OUTPUT_PATH are global variables, which causes every project to output to the same directory. This is a problem if the project is being used as a subproject.

I've replaced these global variables with a per target solution. Now only the targets build by XSC will end up in build/.